+++ iOS/Filmroom/Extension.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import Metal
 
 extension UIViewController {
     
@@ -19,4 +20,32 @@ extension UIViewController {
         present(alertController, animated: true, completion: nil)
     }
     
+}
+
+extension MTLTexture{
+    var toUIImage: UIImage {
+        let bytesPerPixel = 4
+        let imageByteCount = self.width * self.height * bytesPerPixel
+        let bytesPerRow = self.width * bytesPerPixel
+        var src = [UInt8](repeating: 0, count: Int(imageByteCount))
+        
+        let region = MTLRegionMake2D(0, 0, self.width, self.height)
+        self.getBytes(&src, bytesPerRow: bytesPerRow, from: region, mipmapLevel: 0)
+        
+        let bitmapInfo = CGBitmapInfo(rawValue: (CGBitmapInfo.byteOrder32Big.rawValue | CGImageAlphaInfo.premultipliedLast.rawValue))
+        
+        let colorSpace = CGColorSpaceCreateDeviceRGB()
+        let bitsPerComponent = 8
+        let context = CGContext(data: &src,
+                                width: self.width,
+                                height: self.height,
+                                bitsPerComponent: bitsPerComponent,
+                                bytesPerRow: bytesPerRow,
+                                space: colorSpace,
+                                bitmapInfo: bitmapInfo.rawValue)
+        
+        let dstImageFilter = context?.makeImage()
+        
+        return UIImage(cgImage: dstImageFilter!, scale: 0.0, orientation: UIImageOrientation.up)
+    }
 }