+++ iOS/Filmroom/ViewController.swift
@@ -7,6 +7,9 @@
 //
 
 import UIKit
+import Metal
+import MetalKit
+
 
 class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
 
@@ -14,7 +17,8 @@ class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavig
         super.viewDidLoad()
         
         picker.delegate = self
-        
+        device = MTLCreateSystemDefaultDevice()
+        commandQueue = device.makeCommandQueue()
         //CIFilter.registerName("Exposure Filter", constructor: FilterVendor(), classAttributes: [kCIAttributeFilterName: "Exposure Filter"])
         // Do any additional setup after loading the view, typically from a nib.
     }
@@ -35,54 +39,55 @@ class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavig
         
         let exposure = ExposureFilter()
         let expoUnit = ExpSlider.value
-        
+        /*
         let shadow = ShadowFilter()
         let shadowUnit = ShadowSlider.value
         
         let highlight = HighlightFilter()
         let hlUnit = HLSlider.value
-        
-        let contrast = ContrastFilter()
-        let contrastUnit = ContrastSlider.value
+        */
         
         let hsv = HSLFilter()
         let shift = CIVector(x: CGFloat(HueSlider.value), y: CGFloat(SatSlider.value), z: CGFloat(LumSlider.value))
         
+        let contrast = ContrastFilter()
+        let contrastUnit = ContrastSlider.value
+        
         let saturation = SaturationFilter()
-        let satUnit = SatSlider.value
+        let satUnit = SaturationSlider.value
+ 
+        exposure.inputImage = input
+        exposure.inputUnit = CGFloat(expoUnit)
+        /*
+        shadow.inputImage = exposure.outputImage
+        shadow.inputUnit = CGFloat(shadowUnit)
         
-        exposure.setValue(input, forKey: kCIInputImageKey)
-        exposure.setValue(expoUnit, forKey: "inputUnit")
+        highlight.inputImage = shadow.outputImage
+        highlight.inputUnit = CGFloat(hlUnit)
+        */
+        contrast.inputImage = exposure.outputImage
+        contrast.inputUnit = CGFloat(contrastUnit)
         
-        shadow.setValue(exposure.outputImage, forKey: kCIInputImageKey)
-        shadow.setValue(shadowUnit, forKey: "inputUnit")
-        
-        highlight.setValue(shadow.outputImage, forKey: kCIInputImageKey)
-        highlight.setValue(hlUnit, forKey: "inputUnit")
-        
-        contrast.setValue(highlight.outputImage, forKey: kCIInputImageKey)
-        contrast.setValue(contrastUnit, forKey: "inputUnit")
-        
-        saturation.setValue(contrast.outputImage, forKey: kCIInputImageKey)
-        saturation.setValue(satUnit, forKey: "inputUnit")
-        
-        hsv.setValue(saturation.outputImage, forKey: kCIInputImageKey)
-        hsv.setValue(shift, forKey: "inputShift0")
+        saturation.inputImage = contrast.outputImage
+        saturation.inputUnit = CGFloat(satUnit)
+ 
+        hsv.inputImage = saturation.outputImage
+        hsv.inputShift0 = shift
         
         
-        
-        resultImage = hsv.outputImage!
+        resultImage = hsv.outputImage
         
         MyImageView.image = UIImage(ciImage: hsv.outputImage)
     }
 
-    @IBOutlet weak var HueSlider: UISlider!
     @IBOutlet weak var SatSlider: UISlider!
+    @IBOutlet weak var HueSlider: UISlider!
+    @IBOutlet weak var SaturationSlider: UISlider!
     @IBOutlet weak var LumSlider: UISlider!
     @IBOutlet weak var ExpSlider: UISlider!
     @IBOutlet weak var ContrastSlider: UISlider!
-    @IBOutlet weak var HLSlider: UISlider!
-    @IBOutlet weak var ShadowSlider: UISlider!
+//    @IBOutlet weak var HLSlider: UISlider!
+//    @IBOutlet weak var ShadowSlider: UISlider!
     @IBOutlet weak var MyImageView: UIImageView!
     let picker = UIImagePickerController()
     var processedImage: CIImage?
@@ -100,8 +105,6 @@ class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavig
         LumSlider.value = 1
         ExpSlider.value = 0
         ContrastSlider.value = 1
-        HLSlider.value = 0
-        ShadowSlider.value = 0
         SatSlider.value = 1
         
         present(picker, animated: true, completion: nil)
@@ -139,11 +142,185 @@ class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavig
         dismiss(animated: true, completion: nil)
     }
     
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : AnyObject]){
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : Any]) {
         MyImageView.image = info[UIImagePickerControllerOriginalImage] as? UIImage
         processedImage = CIImage(image: MyImageView.image!)
         
         dismiss(animated:true, completion: nil)
+    }
+    
+    var device: MTLDevice!
+    var commandQueue: MTLCommandQueue!
+    var sourceTexture: MTLTexture!
+    
+    
+    var textureLoader: MTKTextureLoader!
+    
+    @IBAction func FFT(_ sender: UIButton) {
+        var commandBuffer = commandQueue.makeCommandBuffer()
+        
+        guard let cgImage = MyImageView.image?.cgImage else {
+            fatalError("Can't open image")
+        }
+        
+        textureLoader = MTKTextureLoader(device: self.device)
+        do {
+            sourceTexture = try textureLoader.newTexture(cgImage: cgImage, options: nil)
+        }
+        catch {
+            fatalError("Can't load texture")
+        }
+        
+        
+        /// A Metal library
+        var defaultLibrary: MTLLibrary!
+        
+        // Load library file
+        do{
+            try defaultLibrary = device.makeDefaultLibrary()
+        }catch{
+            fatalError("Load library error")
+        }
+        
+        // Select library function
+        let reOrderKernel = defaultLibrary.makeFunction(name: "reposition")!
+        
+        // Set pipeline of Computation
+        var pipelineState: MTLComputePipelineState!
+        do{
+            pipelineState = try device.makeComputePipelineState(function: reOrderKernel)
+        }catch{
+            fatalError("Set up failed")
+        }
+        
+        /*
+         * Create new texture for store the pixel data,
+         * or say any data that requires to be processed by the FFT function
+         */
+        let textureDescriptor = MTLTextureDescriptor.texture2DDescriptor(pixelFormat: .rg32Float, width: sourceTexture.width, height: sourceTexture.height, mipmapped: false)
+        var reorderedTexture: MTLTexture!
+        reorderedTexture = self.device.makeTexture(descriptor: textureDescriptor)
+        
+        
+        // config the group number and group size
+        var commandEncoder = commandBuffer?.makeComputeCommandEncoder()
+        
+        /* Figure out the:
+         
+         * The number of threads that are scheduled to execute the same instruction
+         in a compute function at a time.
+         * The largest number of threads that can be in one threadgroup.
+         
+         */
+        let tw = pipelineState.threadExecutionWidth
+        let th = pipelineState.maxTotalThreadsPerThreadgroup / tw
+        
+        let threadPerGroup = MTLSizeMake(tw, th, 1)
+        let threadGroups: MTLSize = MTLSizeMake(Int(self.sourceTexture.width) / threadPerGroup.width, Int(self.sourceTexture.height) / threadPerGroup.height, 1)
+        
+        
+        // Set texture in kernel
+        commandEncoder?.setComputePipelineState(pipelineState)
+        commandEncoder?.setTexture(self.sourceTexture, index: 0)
+        commandEncoder?.setTexture(reorderedTexture, index: 1)
+        
+        // Pass width and length data to GPU
+        var width = self.sourceTexture.width
+        var length = width * self.sourceTexture.height
+        
+        // Set data buffer
+        let bufferW = device.makeBuffer(bytes: &width, length: MemoryLayout<uint>.size, options: MTLResourceOptions.storageModeShared)
+        let bufferL = device.makeBuffer(bytes: &length, length: MemoryLayout<uint>.size, options: MTLResourceOptions.storageModeShared)
+        
+        commandEncoder?.setBuffer(bufferW, offset: 0, index: 0)
+        commandEncoder?.setBuffer(bufferL, offset: 0, index: 1)
+        
+        // Config the thread setting
+        commandEncoder?.dispatchThreadgroups(threadGroups, threadsPerThreadgroup: threadPerGroup)
+        commandEncoder?.endEncoding()
+        
+        // Push the configuration and assignments to GPU
+        commandBuffer?.commit()
+        commandBuffer?.waitUntilCompleted()
+        // Above, finished the re-arrangment
+        
+        // Load function of FFT calculation
+        let fftStageKernel = defaultLibrary.makeFunction(name: "fft_allStage")
+        // Set pipeline of Computation
+        do{
+            pipelineState = try device.makeComputePipelineState(function: fftStageKernel!)
+        }catch{
+            fatalError("Set up failed")
+        }
+        
+        // Calculate the total pixel amount
+        let totalPixel = Float(width * length)
+        
+        // Set texture in kernel
+        for index in 1...Int(log2(totalPixel)){
+            // Start steps of FFT -- Calculate each row
+            // Refresh the command buffer and encoder for each stage
+            commandBuffer = commandQueue.makeCommandBuffer()
+            commandEncoder = commandBuffer?.makeComputeCommandEncoder()
+            
+            
+            commandEncoder?.setComputePipelineState(pipelineState)
+            commandEncoder?.setTexture(reorderedTexture, index: 0)
+            var stage = UInt(index)
+            
+            // Adjust the FFT and complexConjugate to get inverse or complex conjugate
+            var FFT: Int = 1
+            var complexConjugate: Int = 1
+            
+            let bufferS = device.makeBuffer(bytes: &stage, length: MemoryLayout<uint>.size, options: MTLResourceOptions.storageModeShared)
+            let bufferFFT = device.makeBuffer(bytes: &FFT, length: MemoryLayout<Int>.size, options: MTLResourceOptions.storageModeShared)
+            let bufferComplexConjugate = device.makeBuffer(bytes: &complexConjugate, length: MemoryLayout<Int>.size, options: MTLResourceOptions.storageModeShared)
+            
+            commandEncoder?.setBuffer(bufferW, offset: 0, index: 0)
+            commandEncoder?.setBuffer(bufferL, offset: 0, index: 1)
+            commandEncoder?.setBuffer(bufferS, offset: 0, index: 2)
+            commandEncoder?.setBuffer(bufferFFT, offset: 0, index: 3)
+            commandEncoder?.setBuffer(bufferComplexConjugate, offset: 0, index: 4)
+            
+            commandEncoder?.dispatchThreadgroups(threadGroups, threadsPerThreadgroup: threadPerGroup)
+            commandEncoder?.endEncoding()
+            
+            // Push the assignment
+            commandBuffer?.commit()
+            
+        }
+        
+        // Load function of FFT calculation
+        let modulusKernel = defaultLibrary.makeFunction(name: "complexModulus")
+        // Set pipeline of Computation
+        do{
+            pipelineState = try device.makeComputePipelineState(function: modulusKernel!)
+        }catch{
+            fatalError("Set up failed")
+        }
+        
+        
+        let resultDescriptor = MTLTextureDescriptor.texture2DDescriptor(pixelFormat: .rgba8Unorm, width: sourceTexture.width, height: sourceTexture.height, mipmapped: false)
+        var resultTexture: MTLTexture!
+        resultTexture = self.device.makeTexture(descriptor: resultDescriptor)
+        
+        commandBuffer = commandQueue.makeCommandBuffer()
+        commandEncoder = commandBuffer?.makeComputeCommandEncoder()
+        
+        commandEncoder?.setComputePipelineState(pipelineState)
+        commandEncoder?.setTexture(reorderedTexture, index: 0)
+        commandEncoder?.setTexture(resultTexture, index: 1)
+        
+        commandEncoder?.dispatchThreadgroups(threadGroups, threadsPerThreadgroup: threadPerGroup)
+        commandEncoder?.endEncoding()
+        
+        // Push the assignment
+        commandBuffer?.commit()
+        commandBuffer?.waitUntilCompleted()
+        
+        MyImageView.image = resultTexture.toUIImage
+        processedImage = CIImage(image: MyImageView.image!)
+        
     }
 }